Thanks to arion_p for this patch
EpisodeNum / SeriesNum exporting to xmltv format: According to xmltv DTD the series/episode/part number can be in one of two formats xmltv_ns and onscreen (other format may be used but are not defined in xmltv DTD and TVServer does not support them). If no format is specified in the "system" attribute (which is the case with WebEPG), xmltv_ns is assumed.
Now according to the xmltv_ns system the value should be of the form [series].[episode].[parth], where [] denote an optional field. Note that the dots are NOT optional. WebEPG on the other hand exported episode numbers in the format [[part.]episode[.series]]
WebEPG_Episodes_Patch changes WebEPG fixthis