Thanks arion_p for this patch
Epoch time is not handled correctly: If the site timezone is UTC-8 everything works fine, but for any other time zone, programs may be added to the wrong date. I have changed the conversion to epoch date/time to take into account the site timezone. WebEPG_EpochTime_Patch fixes this.