First commit of TechnoTrend refactoring.
Mostly this really is a huge refactoring and tidy up but there are a few important changes too (I think).

- improve API function/enum documentation by trial and error
- add custom tuning implementation
- make the DVB-T aerial power on/off accessible via the SetPowerState() implementation
- add more information to augment the existing driver version logging
- improve "no CA resource" descrambling error handling

There is one notable regression: no automatic timeout handling when the CAM is not ready to descramble. In my opinion that kind of logic should be handled at a higher level as the issue applies to other classes too (eg. Digital Everywhere).