Fixed Mouse Pointer Hiding in Windowed Mode
Updated HandleCursor() to set the MouseTimeOutTime to DateTime.Now() in d3dapp.cs when the mouse pointer is not completly over the form. This prevents the mouse pointer from being hidden when passing over non-client area longer than three seconds and re-entering the form's border or menu bar. Also applies to the context menu of the task bar item when minimized.