Mantis #4129
Added reset of playing times and durations to RemovePlayerProperties() in PropertyManager.cs

A better fix would be that the VideoOverlay doesn't show up unless playback actually has begun. I investigated this and it seems to be too much work for this issue.

In fact it shouldn't come up at all as playback starts in fullscreen. But currently the state is connected to OnVideoChanged() which is the wrong event in general as video dimensions or aspect ratio could be set before the playback can actually begin.

The overlay should only be visible when the first fame is rendered, not before.

With this fix the times are reset when playback it stops. So on subsequent playbacks the old values shouldn't show up.

There is another side effect. The movie title is still the filename and not what is inside the TV/Movie database. This gets updated right before playback starts. Another "cosmetic issue" by design.

The issue is only visible if it takes a second or so before fullscreen playback starts (e.g. MediaInfo takes some time over the network or the file to be processed in larger. 1080 MKVs for example).