Mantis #4188, #4189, #4190, #4191 and #2899
Fixes several glitches in the SkinEngine related to GUIButtonControl.cs usage that relies on rendering methods of GUIFadeLabel.cs

Even though these are independent issues, the fixes play hand in hand and cannot be seperated.

Also did some minor code cleanup in regards to remove inconsistent variable naming and redundant code that does nothing and was actually complicating fixing this issue.

Mantis #2899 should be fixed as a side effect as it's just one symptom of underlying defect.