Clipping for scrolling text didn't work properly when the control's height equals zero
The clipping for scrolling text didn't work properly when the height of the control was set to zero. This happens for labels only. The current constructor does not give them width/height automatically, so it needs to be compensated in the code later.

Added a condition to the clipping. When height of a label is <= 0 the height of the text will be used for setting the clipping region.