Crash when viewing a playlist
Recent commit of fixing the fading of the right most label when viewing a playlist caused a regression.

Reverting the changes and  providing a proper fix.

If the width of a text equals the width of the area it is rendered in, the skin engine fades the text.

This is either intentional or a ">=" instead of a ">" is used somewhere in GUIFadeLabel.cs to decide when fading should be done.

This behaviour is now compensated in GUIPlayListItemListControl.cs by adding +1 to the text width or else the last label would fade.