Fixed switching to already running MP if minimize to Tray
If MP was minimzied to the tray it could not be restored by thread messages, because the controls were hidden. Hidden control mean, that MP does not show up in the ALT-TAB task switcher. As MP is no background worker application this behavior was inconsitent.

In addition to WM_SHOWWINDOW an WM_ACTIVATE message is send that reactivates MP in case it is minized to the tray.

Furthermode the check if MediaPortal was already running was too late in the code actually resulting in creation of new logs files. Moved the check further up in the start sequence.