Clean card/tuner/device inheritance structure by introducing DeviceDirectShow. The DVB-IP/IPTV and B2C2 (SS2) classes now don't need to depend on TvCardDvbBase (which is really a BDA base class). The HD-PVR/Colossus and analog classes can now reuse more code as support is DirectShow based.
As a side-effect of the above, EPG grabbing will no longer work with DVB-IP and B2C2 tuners. I don't think IPTV tuners matter as they likely don't have EPG data in the stream. B2C2 tuners... well, it matters... but lets clean up the EPG grabbing code and implement it in such a way that it can be shared without all the awful hacks!

Also removed add/remove of IPTV source filter on tune. georgius confirmed that this is not necessary with the new filter. At this point I don't really care if I break IPTV support or the Elecard source is not compatible - these sorts of hacks need to be removed.

I also took the opportunity to clean up repeated use of "if (hr != 0)" in the B2C2 class. There is loads more of this, but one step at a time.