Make sure m_Streams isn't null in TimeStretchFilter before using it
When LAV Audio is configured for bit streaming, but the audio device
does not support it a PCM type will be negotiated (which partially
enables the TimeStretchFilter). LAV Audio will then attempt a dynamic
format change to a bit streaming type and the TSF causes a NPE when it
tries to cleanup.