Treat device period in settings as a default value
So don't set it in InitAudioClient. Some filters (cough LAV Audio) test
multiple formats at a time and choose the one they want. When device
period is set it changes it permanently for the instance. This can cause
problems if a bit streamable format is tested (which lowers latency),
but a PCM type is selected after it (or dynamically changed later).