Send EC_STARVATION event only when filter is in paused state
WASAPI render's buffer won't be ever 100% full and if player queries the graph
state it could trigger the graph to be paused as rendered is considered to be
buffering.

For example GraphStudioNext will query the graph state when user opens filter's
context menu and this will cause unwanted pause and resume cycle to be done.