Update LIVE555 to version 2015.07.31. TsReader modifications are currently left out because I don't want to have to merge TsReader changes with master. This means TsReader won't build in this branch from now on.
We have one modification in GenericMediaServer.hh:
noteLiveness() must be virtual so that we can override the default behaviour. My understanding - and I could be wrong - is that when paused, TsReader doesn't do anything to let the streaming server know that the client is still alive. Therefore, under normal circumstances LIVE555 on the server side would drop the session after ~65 seconds. We don't want that. Hence the modification.