Add a "transmitter" field to each scan section to make it possible to scan a single transmitter/tuning detail - standard ***OR*** full NIT or fast NIT info - *without having to manually specify all the tuning parameters* (which many people find too hard). This required some re-jigging of the advanced options for DVB-C, satellite and DVB-T/T2 scanning. In particular, we now have: - the "use manual tuning" check box - the "manual tuning" group box - "predefined provider" scan type replaced with "standard"
The changes to CardDvbS also include integration of TunerSatellite, so that the "satellite" field only lists the satellites associated with the tuner.

Note that the upper fields related to tuning detail selection and other fields that depend on config (eg. broadcast standard) are now only updated on section activation *when a scan is not in progress*. This ensures that the selected options don't change while a scan is happening.