Refactor analog tuner graph building yet again!
The overall goal is to have the components be more self-contained and loosen their coupling. In particular, the SetAudioCapture() and SetStreamConfigInterface() capture component functions were bothering me. The capture component should have been capable of finding it's own components and interfaces... and now it can. This required quite significant reworking of the output pin identification code.

Pin-related code is now more generic, and this enabled it to be extracted to the new base component class. It's also slightly more efficient compared to the former encoder FindPinByCategoryOrMediaType() function, because it enumerates the pins only once per filter rather than once per target pin type.

Some small enhancements to the encoder VBI filter handling...
On XP, use the tee/sink-to-sink filter to connect both the WST codec and CC decoder when possible. (Previously we only connected the WST codec.)
The previous Hauppauge SAA7164 VBI bug workaround wouldn't have worked. Testing with GraphStudio showed that the graph failed to start when the tee/sink-to-sink filter input was not connected. Therefore we have to actually remove and re-add the VBI filters, not just disconnect and reconnect the VBI splitter input.

I've also moved default analog tuner configuration code to the relevant component where possible.