Unfortunately all existing section decoder implementations assume that they'll receive sections in the order that they were broadcast. When you introduce more than one processor (as I have previously done, to try to improve performance), you also introduce the possibility that sections will be delivered out of order to decoders.
The MHW parser is particularly fussy with respect to section delivery order:
1. Version 2 description multi-section table sections must be received in order and combined before they can be parsed.
2. Change/completion detection for event and description tables/segments are based on identifier repetition/change or wrap around mechanisms which heavily rely on in-order section delivery.

This commit adds "requireSequentialDispatch logic" as a workaround for the MHW parser issues. However, other section decoders are potentially affected by similar problems. For example, many decoders detect change via the section version number. Out-of-order delivery will introduce jitter in those changes. That may or may not be a problem.