Ensure device index selection includes *all* USB tuners.
This change only affects CI support.

I'm not certain of this, but TBS may have stopped using the QBOX brand for newer tuners. If that's the case, this change will only affect systems with more than one TBS tuner with CI support, where at least one of them has a "QBOX..." name.

Please note the oldest USB tuner with CI slot I'm aware of is the TBS 5980, and that tuner is named "TBS...". Therefore this change is _likely_ to have no effect.